+++ packages/bluesnap-direct-integration/src/index.ts
@@ -1,0 +1,1 @@
+export { default as createBlueSnapDirectCreditCardPaymentStrategy } from './create-bluesnap-direct-credit-card-payment-strategy';