+++ packages/bluesnap-direct-integration/src/mocks/bluesnap-direct-method.mock.ts
@@ -1,0 +1,18 @@
+import { PaymentMethod } from '@bigcommerce/checkout-sdk/payment-integration-api';
+
+export function getBlueSnapDirect(): PaymentMethod {
+    return {
+        clientToken: 'pfToken',
+        config: {
+            cardCode: true,
+            displayName: 'Credit Card',
+            isHostedFormEnabled: true,
+            testMode: true,
+        },
+        gateway: 'bluesnapdirect',
+        id: 'credit_card',
+        method: 'credit_card',
+        supportedCards: ['AMEX', 'CUP', 'DINERS', 'DISCOVER', 'JCB', 'MC', 'VISA'],
+        type: 'PAYMENT_TYPE_API',
+    };
+}