+++ packages/core/src/payment/payment-request-options.ts
@@ -19,7 +19,7 @@ import { MolliePaymentInitializeOptions } from './strategies/mollie';
 import { MonerisPaymentInitializeOptions } from './strategies/moneris';
 import { OpyPaymentInitializeOptions } from './strategies/opy';
 import { PaypalExpressPaymentInitializeOptions } from './strategies/paypal';
-import { PaypalCommerceInitializeOptions } from './strategies/paypal-commerce';
+import { PaypalCommercePaymentInitializeOptions } from './strategies/paypal-commerce';
 import { SquarePaymentInitializeOptions } from './strategies/square';
 import { StripeUPEPaymentInitializeOptions } from './strategies/stripe-upe';
 import { StripeV3PaymentInitializeOptions } from './strategies/stripev3';
@@ -135,7 +135,7 @@ export interface BasePaymentInitializeOptions extends PaymentRequestOptions {
      * The options that are required to initialize the PayPal Commerce payment method.
      * They can be omitted unless you need to support PayPal Commerce.
      */
-    paypalcommerce?: PaypalCommerceInitializeOptions;
+    paypalcommerce?: PaypalCommercePaymentInitializeOptions;
 
     /**
      * The options that are required to initialize the Square payment method.