+++ packages/payment-integration-api/src/config/config.ts
@@ -88,9 +88,10 @@ export interface StoreCurrency {
     thousandsSeparator: string;
 }
 
-export type UserExperienceSettingNames = 'walletButtonsOnTop';
-
-export type UserExperienceSettings = { [key in UserExperienceSettingNames]: boolean };
+export interface UserExperienceSettings {
+    walletButtonsOnTop: boolean;
+    floatingLabelEnabled: boolean;
+}
 
 export interface CheckoutSettings {
     features: { [featureName: string]: boolean };